requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# OCR Libraries
easyocr>=1.7.0
paddleocr>=2.7.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# AI/ML
groq>=0.4.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20
scikit-learn>=1.3.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.0.0

# Development (optional)
jupyter>=1.0.0
notebook>=7.0.0
